+++ src/main/java/seedu/address/logic/commands/StatusDeleteCommand.java
@@ -21,6 +21,7 @@ public class StatusDeleteCommand extends Command {
     public static final String MESSAGE_SUCCESS = "Deleted status: %1$s";
     public static final String MESSAGE_STATUS_NOT_FOUND = "The specified status does not exist in the catalog";
     public static final String MESSAGE_CANNOT_DELETE_DEFAULT_STATUS = "Cannot delete default status: Unknown";
+    public static final String MESSAGE_STATUS_IN_USE = "Cannot delete status currently assigned to persons";
 
     private final Status toDelete;
 
@@ -42,6 +43,10 @@ public class StatusDeleteCommand extends Command {
 
         if (!model.hasStatus(toDelete)) {
             throw new CommandException(MESSAGE_STATUS_NOT_FOUND);
+        }
+
+        if (!model.getPersonsMatching(person -> person.getStatus().equals(toDelete)).isEmpty()) {
+            throw new CommandException(MESSAGE_STATUS_IN_USE);
         }
 
         model.deleteStatus(toDelete);
@@ -69,4 +74,3 @@ public class StatusDeleteCommand extends Command {
                 .toString();
     }
 }
-
