+++ src/main/java/seedu/address/logic/commands/TeamDeleteCommand.java
@@ -21,6 +21,7 @@ public class TeamDeleteCommand extends Command {
     public static final String MESSAGE_SUCCESS = "Deleted team: %1$s";
     public static final String MESSAGE_TEAM_NOT_FOUND = "The specified team does not exist in the catalog";
     public static final String MESSAGE_CANNOT_DELETE_DEFAULT_TEAM = "Cannot delete default team: Unassigned Team";
+    public static final String MESSAGE_TEAM_IN_USE = "Cannot delete team currently assigned to persons";
 
     private final Team toDelete;
 
@@ -42,6 +43,10 @@ public class TeamDeleteCommand extends Command {
 
         if (!model.hasTeam(toDelete)) {
             throw new CommandException(MESSAGE_TEAM_NOT_FOUND);
+        }
+
+        if (!model.getPersonsMatching(person -> person.getTeam().equals(toDelete)).isEmpty()) {
+            throw new CommandException(MESSAGE_TEAM_IN_USE);
         }
 
         model.deleteTeam(toDelete);
@@ -69,4 +74,3 @@ public class TeamDeleteCommand extends Command {
                 .toString();
     }
 }
-
